requirements.txt
Flask==3.0.0
imread_from_url==0.1.3
numpy==1.24.1
onnxruntime==1.16.0
opencv_python==4.7.0.68
Pillow==10.0.1
